bugfix cant serialize json with characters >127 signed char -> int problem

+++ lang/c++/impl/json/JsonIO.hh
@@ -193,7 +193,7 @@ class AVRO_DECL JsonGenerator {
                 escape('t', b, p);
                 break;
             default:
-                if (! iscntrl(*p)) {
+                if (! iscntrl((uint8_t) *p)) {
                     continue;
                 }
                 write(b, p);
